[engine][server] Refactoring. Moved FieldService to engine.
