+++ CodingDojo/server/src/main/java/com/codenjoy/dojo/services/multiplayer/Spreader.java
@@ -23,7 +23,6 @@ package com.codenjoy.dojo.services.multiplayer;
  */
 
 import com.codenjoy.dojo.services.Deal;
-import com.codenjoy.dojo.services.FieldService;
 import com.codenjoy.dojo.services.Player;
 import com.google.common.collect.LinkedHashMultimap;
 import com.google.common.collect.Multimap;
