+++ CodingDojo/server/src/test/java/com/codenjoy/dojo/services/controller/chat/ChatControllerTest.java
@@ -24,7 +24,6 @@ package com.codenjoy.dojo.services.controller.chat;
 
 
 import com.codenjoy.dojo.services.Deal;
-import com.codenjoy.dojo.services.FieldService;
 import com.codenjoy.dojo.services.PlayerService;
 import com.codenjoy.dojo.services.chat.ChatAuthority;
 import com.codenjoy.dojo.services.chat.ChatService;
@@ -34,7 +33,7 @@ import com.codenjoy.dojo.services.controller.AbstractControllerTest;
 import com.codenjoy.dojo.services.controller.Controller;
 import com.codenjoy.dojo.services.dao.Chat;
 import com.codenjoy.dojo.services.helper.Helpers;
-import org.json.JSONObject;
+import com.codenjoy.dojo.services.multiplayer.FieldService;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.stubbing.Answer;
