+++ CodingDojo/server/src/test/java/com/codenjoy/dojo/services/helper/LoginHelper.java
@@ -25,6 +25,7 @@ package com.codenjoy.dojo.services.helper;
 import com.codenjoy.dojo.services.*;
 import com.codenjoy.dojo.services.dao.Registration;
 import com.codenjoy.dojo.services.hash.Hash;
+import com.codenjoy.dojo.services.multiplayer.FieldService;
 import com.codenjoy.dojo.services.nullobj.NullDeal;
 import com.codenjoy.dojo.services.nullobj.NullPlayer;
 import com.codenjoy.dojo.services.security.GameAuthorities;
