+++ CodingDojo/server/src/test/java/com/codenjoy/dojo/web/rest/AbstractRestControllerTest.java
@@ -31,6 +31,7 @@ import com.codenjoy.dojo.services.dao.Chat;
 import com.codenjoy.dojo.services.dao.Registration;
 import com.codenjoy.dojo.services.helper.Helpers;
 import com.codenjoy.dojo.services.log.DebugService;
+import com.codenjoy.dojo.services.multiplayer.FieldService;
 import com.codenjoy.dojo.services.room.RoomService;
 import com.codenjoy.dojo.services.semifinal.SemifinalService;
 import com.codenjoy.dojo.utils.smart.SmartAssert;
